Code clean up and restructing
This commit contains a large number of changes that can be summarized
in the following 3 categories:

1) Introduces a new cache coherence mechanism

2) Clean-up sampling methods

3) Add support for power model for in-order core

+++ misc/libsuc/GStats.cpp
@@ -122,7 +122,7 @@ GStatsCntr::GStatsCntr(const char *format,...)
 
 double GStatsCntr::getDouble() const
 {
-  return (double)data;
+  return data;
 }
 
 void GStatsCntr::reportValue() const
@@ -132,7 +132,7 @@ void GStatsCntr::reportValue() const
 
 int64_t GStatsCntr::getSamples() const 
 { 
-  return data;
+  return (int64_t)data;
 }
 
 /*********************** GStatsAvg */
